Added initial boolean values to page1
